# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SilentRelay API
  description: |
    Secure end-to-end encrypted messaging API for SilentRelay.
    
    ## Authentication
    Most endpoints require JWT authentication via `Authorization: Bearer <token>` header.
    
    ## Rate Limiting
    - SMS endpoints: 5 requests/minute/IP
    - Auth endpoints: 10 requests/minute/IP  
    - Protected endpoints: 100 requests/minute/user
    
    ## Security
    All communications use end-to-end encryption. The server never has access to message content.
  version: 1.0.0
  contact:
    name: SilentRelay Support
    url: https://silentrelay.com.au
  license:
    name: Proprietary
    url: https://silentrelay.com.au/terms
servers:
  - url: https://silentrelay.com.au/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or registration
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone_number:
          type: string
          example: "+1234567890"
        username:
          type: string
          example: "john_doe"
        display_name:
          type: string
          example: "John Doe"
        avatar_url:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        created_at:
          type: string
          format: date-time
    Device:
      type: object
      properties:
        device_id:
          type: string
          format: uuid
        device_name:
          type: string
          example: "Chrome on MacBook"
        device_type:
          type: string
          enum: [web, mobile, desktop]
        is_primary:
          type: boolean
        last_active:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
paths:
...
  /users/me/prekeys:
    post:
      tags:
        - Users
      summary: Upload prekeys
      description: Uploads one-time prekeys for E2EE key exchange
      operationId: uploadPrekeys
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prekeys
              properties:
                prekeys:
                  type: array
                  items:
                    type: object
                    properties:
                      key_id:
                        type: integer
                      public_key:
                        type: string
                        format: byte
      responses:
        '200':
          description: Prekeys uploaded
        '401':
          description: Unauthorized
  /users/{userId}/keys:
    get:
      tags:
        - Users
      summary: Get user's public keys
      description: Retrieves a user's public keys for establishing E2EE session
      operationId: getUserKeys
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User's public keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  identity_key:
                    type: string
                    format: byte
                  signed_prekey:
                    type: string
                    format: byte
                  signed_prekey_signature:
                    type: string
                    format: byte
                  one_time_prekey:
                    type: string
                    format: byte
                    description: Optional one-time prekey if available
        '404':
          description: User not found
...
  /users/check-username/{username}:
    get:
      tags:
        - Users
      summary: Check username availability
      operationId: checkUsername
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Username check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
  /users/search:
    get:
      tags:
        - Users
      summary: Search users
      description: Search for users by username or phone number
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 3
          description: Search query (minimum 3 characters)
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'

  # =====================
  # DEVICES
  # =====================
  /devices:
    get:
      tags:
        - Devices
      summary: List devices
      description: Returns all devices linked to the user's account
      operationId: getDevices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
  /devices/{deviceId}:
    delete:
      tags:
        - Devices
      summary: Remove device
      description: Unlinks a device from the user's account
      operationId: removeDevice
      security:
        - bearerAuth: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device removed
        '403':
          description: Cannot remove primary device
  /devices/{deviceId}/primary:
    put:
      tags:
        - Devices
      summary: Set primary device
      description: Sets a device as the primary device
      operationId: setPrimaryDevice
      security:
        - bearerAuth: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Primary device updated

  # =====================
  # DEVICE APPROVAL
  # =====================
  /device-approval/request:
    post:
      tags:
        - Device Approval
      summary: Request device approval
      description: Initiates the device linking process for a new device
      operationId: requestDeviceApproval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone_number
                - device_id
                - device_name
                - device_type
              properties:
                phone_number:
                  type: string
                device_id:
                  type: string
                  format: uuid
                device_name:
                  type: string
                device_type:
                  type: string
                  enum: [web, mobile, desktop]
                public_device_key:
                  type: string
                  format: byte
      responses:
        '200':
          description: Approval request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
                  verification_code:
                    type: string
                    description: 6-digit code to display on new device
                  expires_at:
                    type: string
                    format: date-time
  /device-approval/verify:
    post:
      tags:
        - Device Approval
      summary: Verify approval code
      description: Verifies the approval code entered on the new device
      operationId: verifyApprovalCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - request_id
                - code
              properties:
                request_id:
                  type: string
                  format: uuid
                code:
                  type: string
      responses:
        '200':
          description: Code verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  status:
                    type: string
                    enum: [pending, approved, denied, expired]
  /device-approval/{requestId}/status:
    get:
      tags:
        - Device Approval
      summary: Check approval status
      description: Polls for the approval status of a device linking request
      operationId: checkApprovalStatus
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Approval status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, approved, denied, expired]
                  access_token:
                    type: string
                    description: Present only when status is approved
                  refresh_token:
                    type: string
  /device-approval/pending:
    get:
      tags:
        - Device Approval
      summary: Get pending approvals
      description: Returns pending device approval requests for the authenticated user
      operationId: getPendingApprovals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending approvals
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      type: object
                      properties:
                        request_id:
                          type: string
                          format: uuid
                        device_name:
                          type: string
                        device_type:
                          type: string
                        created_at:
                          type: string
                          format: date-time
  /device-approval/{requestId}/approve:
    post:
      tags:
        - Device Approval
      summary: Approve device
      description: Approves a device linking request from the primary device
      operationId: approveDevice
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device approved
  /device-approval/{requestId}/deny:
    post:
      tags:
        - Device Approval
      summary: Deny device
      description: Denies a device linking request
      operationId: denyDevice
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device denied

  # =====================
  # MESSAGES
  # =====================
...
  /pin:
    get:
      tags:
        - PIN
      summary: Get PIN status
      description: Check if user has a PIN set (returns encrypted PIN hash)
      operationId: getPIN
      security:
        - bearerAuth: []
      responses:
        '200':
          description: PIN status
          content:
            application/json:
              schema:
                type: object
                properties:
                  has_pin:
                    type: boolean
                  encrypted_pin:
                    type: string
                    format: byte
                    description: Present only if has_pin is true

    post:
      tags:
        - PIN
      summary: Set PIN
      operationId: setPIN
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - encrypted_pin
              properties:
                encrypted_pin:
                  type: string
                  format: byte
                  description: Client-side encrypted PIN hash
      responses:
        '200':
          description: PIN set

    delete:
      tags:
        - PIN
      summary: Delete PIN
      operationId: deletePIN
      security:
        - bearerAuth: []
      responses:
        '200':
          description: PIN deleted

  # =====================
  # WEBRTC
  # =====================
...